requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
splunk-sdk>=1.7.0
python-dotenv>=1.0.0
modelcontext>=0.1.0 # Ensure you have the correct/latest version